Add new widgets for style selection and takes management
- Implemented StylePicker for waveform style selection with live previews.
- Created TakesView to display recordings with mini waveforms and context menus.
- Introduced TransportBar for playback controls and time display.
- Added tests for application functionality, models, stream handling, and example scripts.
- Enhanced AppearanceModel and PresetStore for better appearance management.
- Improved StreamSender and StreamReceiver for robust audio streaming.

+++ src/audiowave/audio/player.py
@@ -8,9 +8,9 @@ import numpy as np
 from PySide6.QtCore import QObject, QTimer, Signal
 from PySide6.QtMultimedia import QAudio, QAudioDevice, QAudioFormat, QAudioSink, QMediaDevices
 
+from audiowave.core.annotations import Loop
 from audiowave.core.clip import AudioClip
 from audiowave.core.format import SampleFormat, encode_pcm
-from audiowave.core.annotations import Loop
 
 from .pcm_source import PcmSource
 
@@ -218,7 +218,7 @@ class AudioPlayer(QObject):
             return False
 
         clip = self._clip
-        self._session_rate = max(int(round(clip.sample_rate * self._speed)), 1)
+        self._session_rate = max(round(clip.sample_rate * self._speed), 1)
         fmt = QAudioFormat()
         fmt.setSampleRate(self._session_rate)
         fmt.setChannelCount(clip.channels)
